+++ python/src/jazzy_fish/resources/01234_f233650/adjective.txt
