+++ python/src/jazzy_fish/resources/012_8562fb9/adjective.txt
