+++ python/src/jazzy_fish/resources/024_84f184f/verb.txt
