+++ python/tests/test_decode.py
@@ -6,7 +6,7 @@ sys.path.insert(
     0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "src"))
 )
 
-from encoder.encoder import WordEncoder, EncoderException, Wordlist
+from jazzy_fish.encoder import WordEncoder, EncoderException, Wordlist
 
 
 class TestDecode(unittest.TestCase):
