+++ python/tests/test_generator.py
@@ -10,7 +10,7 @@ sys.path.insert(
     0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "src"))
 )
 
-from encoder.generator import (
+from jazzy_fish.generator import (
     Generator,
     GeneratorException,
     Resolution,
