+++ python/tests/test_select.py
@@ -6,7 +6,7 @@ sys.path.insert(
     0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "src"))
 )
 
-from preprocessor.helpers import least_similar_words
+from jazzy_fish_tools.helpers import least_similar_words
 
 
 class TestSelect(unittest.TestCase):
